requirements.txt
fastapi==0.118.0
httpx==0.28.1
pydantic==2.11.9
python-dotenv==1.1.1
python-multipart==0.0.20
SQLAlchemy==2.0.43
uvicorn==0.37.0
tzdata
Jinja2==3.1.6
